[Fix] 버그 수정 및 강화 이펙트 변경 및 추가
작업 내역
1. 게임 재접속시 보유 세트 카운트가 초기화되는 버그 수정
2. 메인 과제 csv파일 교체
3. 메인과제 클리어시 다음 과제의 보상을 지급하던 버그 수정
4. 돈 관련 변수 int -> long으로 변경(최대 저장값이 약21억에서 922경으로 변경) (데이터 베이스 돈 관련 값은 변경하지 않았으므로, 21억이 넘어가지 않도록 해야함)
5. 레시피 업그레이드시, "직원 업그레이드"문구로 나오던 문제 수정
6. 손님 음식 섭취시 들고있는 음식이 지나가는 손님 레이어 앞에 존재하는 문제 수정
7. 배경 소리 0.7 -> 0.9로 변경
8. 업그레이드 사운드 추가
9. 스탭, 음식, 뽑기 아이템 강화 이펙트 변경 및 추가

+++ Assets/Scripts/Restaurant/DropCoinArea.cs
@@ -19,8 +19,8 @@ public class DropCoinArea : MonoBehaviour
     [SerializeField] private AudioClip _dropCoinSound;
 
     private List<PointerDownSpriteRenderer> _coinList = new List<PointerDownSpriteRenderer>();
-    private int _currentMoney;
-    public int CurrentMoney => _currentMoney;
+    private long _currentMoney;
+    public long CurrentMoney => _currentMoney;
 
     private int _currentCoinCount;
     public int Count => _currentCoinCount;
@@ -43,7 +43,7 @@ public class DropCoinArea : MonoBehaviour
     }
 
 
-    public void LoadData(int coinCount, int money)
+    public void LoadData(int coinCount, long money)
     {
         _currentCoinCount = Mathf.Clamp(coinCount, 0, _maxCoinCount);
         _currentMoney = money;
@@ -105,7 +105,7 @@ public class DropCoinArea : MonoBehaviour
             return;
 
         float endTime = _coinEndTime;
-        int currentMoney = _currentMoney;
+        long currentMoney = _currentMoney;
         _currentMoney = 0;
         _currentCoinCount = 0;
         SoundManager.Instance.PlayEffectAudio(SoundEffectType.GoldSound);
@@ -134,7 +134,7 @@ public class DropCoinArea : MonoBehaviour
             return;
 
         float endTime = _coinEndTime;
-        int currentMoney = _currentMoney;
+        long currentMoney = _currentMoney;
         int currentCoinCount = _currentCoinCount;
         _currentMoney = 0;
         _currentCoinCount = 0;
